Final Changes Commit  - ENP-232

+++ NPD/EmcureNPD.Web/Models/ProjectViewModel.cs
@@ -75,13 +75,13 @@ namespace EmcureNPD.Web.Models
 
     public class CumulativePhaseWiseBudgetView
     {
-        public float Feasability { get; set; }
-        public float Prototype { get; set; }
-        public float ScaleUp { get; set; }
-        public float AMV { get; set; }
-        public float Exhibit { get; set; }
-        public float Filing { get; set; }
-        public float Total { get; set; }
+        public double Feasability { get; set; }
+        public double Prototype { get; set; }
+        public double ScaleUp { get; set; }
+        public double AMV { get; set; }
+        public double Exhibit { get; set; }
+        public double Filing { get; set; }
+        public double Total { get; set; }
     }
     public class AdditionalCostView
     {
